[battery_reminder] make BatteryDisplay lib in lib directory
