+++ firmware/atom_s3_rosserial_message_display/platformio.ini
@@ -10,6 +10,7 @@ lib_deps =
     iory/i2c-for-esp32 @ 0.4.1
     mathertel/OneButton
     lovyan03/LovyanGFX@^1.1.9
+    BatteryDisplay
 build_flags =
 	-DARDUINO_USB_CDC_ON_BOOT=1
 	-DARDUINO_USB_MODE=1
